requirements.txt
opencv-python 
mediapipe 
face_recognition 
faiss-cpu 
numpy
pyheif
matplotlib
pillow 
pillow-heif
dlib
mtcnn
aiofiles

